Port VS unittests to gtest
