+++ TestFiles/CPP_UnitTests/3mfbase10_box.3mf
