+++ TestFiles/CPP_UnitTests/3mfbase12_holed_cube.3mf
