+++ TestFiles/CPP_UnitTests/3mfbase1_cube.3mf
