+++ TestFiles/CPP_UnitTests/3mfbase4_Mesh1.3mf
