+++ TestFiles/CPP_UnitTests/3mfbase8_VertexTet1.3mf
