+++ TestFiles/CPP_UnitTests/3mfbase9_ladybug.3mf
