+++ TestFiles/CPP_UnitTests/Texture_093.3mf
