+++ UnitTests/CPP/Include/UnitTest_Utilities.h
@@ -1,0 +1,49 @@
+/*++
+
+Copyright (C) 2018 3MF Consortium
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without modification,
+are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright notice,
+this list of conditions and the following disclaimer in the documentation
+and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+Abstract:
+
+UnitTest_Utilities.h: Utilities for UnitTests
+
+--*/
+
+#ifndef __NMR_UNITTEST_UTILITIES
+#define __NMR_UNITTEST_UTILITIES
+
+#include "gtest/gtest.h"
+
+#ifdef LTESTFILESPATH
+const std::wstring sTestFilesPath = LTESTFILESPATH;
+#else
+const std::wstring sTestFilesPath = L"TestFiles";
+#endif
+
+#ifdef LOUTFILESPATH
+const std::wstring sOutFilesPath = LOUTFILESPATH;
+#else
+const std::wstring sOutFilesPath = L"..\TestOutput";
+#endif
+
+#endif //__NMR_UNITTEST_UTILITIES
