+++ UnitTests/C_Interface/Source/UnitTest_Attachments.cpp
@@ -30,7 +30,7 @@ in 3MF files.
 
 --*/
 
-#include "UnitTests/UnitTest_Utilities.h"
+#include "UnitTest_Utilities.h"
 #include "Model/COM/NMR_DLLInterfaces.h"
 
 #include <vector>
