+++ UnitTests/C_Interface/Source/UnitTest_Production.cpp
@@ -30,7 +30,7 @@ Defines Unittests for reading and writing the production extenstion to 3MF.
 
 --*/
 
-#include "UnitTests/UnitTest_Utilities.h"
+#include "UnitTest_Utilities.h"
 #include "Model/COM/NMR_DLLInterfaces.h"
 
 #include <vector>
